requirements.txt
pyttsx3==2.90
PyYAML==6.0
pygame==2.1.2


